feat: ターミナルUIの改修
- CMakeLists.txt: mg_tuiパッケージのインストール設定を追加
- __init__.py: AppState、RosBackend、SystemClient、MgTuiAppをインポート
- app.py: MgTuiAppクラスを実装し、UIの構成と更新ロジックを追加
- ros_backend.py: RosBackendクラスを実装し、ROSトピックの購読とサービス呼び出しを追加
- state.py: AppStateデータクラスを定義し、状態管理のための変数を追加
- system_client.py: SystemClientクラスを実装し、サーバーからの状態ポーリング機能を追加
- tui_node.py: エントリーポイントを整理し、MgTuiAppとRosBackendの初期化を追加
