+++ mg_ui/mg_tui/CMakeLists.txt
@@ -5,6 +5,8 @@ find_package(ament_cmake REQUIRED)
 find_package(ament_cmake_python REQUIRED)
 find_package(rclpy REQUIRED)
 
+ament_python_install_package(mg_tui)
+
 install(DIRECTORY launch
   DESTINATION share/${PROJECT_NAME}
 )
